refactor : component import 간소화
