+++ src/components/IssueItem.tsx
@@ -1,6 +1,7 @@
 import { AiOutlineComment } from 'react-icons/ai';
 import { Link } from 'react-router-dom';
 import { styled } from 'styled-components';
+
 interface IssueItemProps {
   issue: IssueType;
 }
