+++ src/components/IssueList.tsx
@@ -1,27 +1,26 @@
+import { AdBanner, Error, IssueItem, Loading } from '.';
 import { useRef } from 'react';
 import { styled } from 'styled-components';
 
 import { getIssues } from '@/api/issues';
 import { useIssuesDispatch, useIssuesState } from '@/context/issuesContext';
-import { useIntersectionObserver } from '@/hook/customhook';
-import AdBanner from './AdBanner';
-import Error from './Error';
-import IssueItem from './IssueItem';
-import Loading from './Loading';
+import { useIntersectionObserver } from '@/hook/observer';
 
 const IssueList = () => {
   const state = useIssuesState();
   const dispatch = useIssuesDispatch();
   const count = useRef(1);
   const { data, loading, error } = state.issues;
+  const setObservationTarget = useIntersectionObserver(fetchIssues);
+
   const list = data?.map((issue: IssueType, i: number) => {
-    return i > 1 && i % 5 == 0 ? <AdBanner key={i} /> : <IssueItem key={i} issue={issue} />;
+    return i > 1 && i % 5 == 4 ? <AdBanner key={i} /> : <IssueItem key={i} issue={issue} />;
   });
-  const fetchIssues = () => {
+
+  function fetchIssues() {
     getIssues(dispatch, count.current);
     count.current++;
-  };
-  const setObservationTarget = useIntersectionObserver(fetchIssues);
+  }
 
   return (
     <ListBox>
@@ -39,4 +38,5 @@ const ListBox = styled.div`
   gap: 20px;
   padding-bottom: 50px;
 `;
+
 export default IssueList;
