+++ src/components/index.ts
@@ -1,0 +1,5 @@
+export { default as AdBanner } from './AdBanner';
+export { default as Error } from './Error';
+export { default as IssueItem } from './IssueItem';
+export { default as IssueList } from './IssueList';
+export { default as Loading } from './Loading';
